runner: split gluster cli as one separate storage handler
This idea comes from the glusterfs project's cli code.

Usage:
 gluster help		-display help for gluster commands
 ceph help		-display help for ceph commands, TODO
 global help		-display help for global commands, TODO
 version		-display the version of nbd-cli

Usage:

	gluster help
		display help for gluster commands
	gluster create <volname@host:/path> [prealloc] <size SIZE> <host HOST>
		create path file on the volname volume, prealloc is false as default, and the SIZE is valid with B, K(iB), M(iB), G(iB), T(iB), P(iB), E(iB), Z(iB), Y(iB)
	gluster delete <volname@host:/path> <host HOST>
		delete path file on the volname volume
	gluster map <volname@host:/path> [nbd-device] [timeout TIME] <host HOST> [readonly]
		map path file to the nbd device, as default the timeout 0, none readonly
	gluster unmap <nbd-device>
		unmap the nbd device
	gluster list <map|unmap|all>
		list the mapped|unmapped|all nbd devices, all as default
